Client: Format instance results without state
The v2 response may sometimes return a successful response for updating
(starting) an instance but with missing 'state' subfields. This was also
the only field for which we did not check existence before using them.
Fall back to a successful 'running' state if it is not given, like v1
does (due to lack of information in the response).